Adapting the EncrytionManager test to work both on MacOS and Linux platforms
